architecture re-work
