--- src/QuicKart-Backend/Aisle.js
@@ -1,3 +1,0 @@
-// Model the data for an aisle here.
-// Look into using realm (MongoDB tech) https://www.mongodb.com/docs/realm/sdk/react-native/
-// Write code to write aisle objects to database
