--- src/QuicKart-Backend/Bay.js
@@ -1,3 +1,0 @@
-// Model the data for a bay here.
-// Look into using realm (MongoDB tech) https://www.mongodb.com/docs/realm/sdk/react-native/
-// Write code to write bay objects to database
