--- src/QuicKart-Frontend/screens/ShoppingList.js
@@ -1,3 +1,0 @@
-// See wire frame at https://www.figma.com/file/KEA3dYb7dAagexyDMAWK4a/Shopping-List-Wireframe
-// This screen needs to show the shopping list title, items search bar, add button, display list
-// of items, total items, and start shopping button.
