--- src/QuicKart-Frontend/screens/ShoppingRoute.js
@@ -1,3 +1,0 @@
-// See wire frame at (place figma file URL here for displaying cards with item locations)
-// This screen will display the route to the user card by card/item by item.
-// Also needs to display the shopping list with items being checked off as the user collects items.
