--- src/QuicKart-Frontend/screens/Welcome.js
@@ -1,3 +1,0 @@
-// See wire frame at https://www.figma.com/file/VPZijQXuBe1PuWynL63ajf/Welcome-Page-Wireframe
-// This screen will display the welcome page with welcome text, short app intro
-// (bulleted list of things they can do within the app), Create List button.
