--- src/QuicKart-Frontend/styles/shared.css
@@ -1,2 +1,0 @@
-/* This is where any shared styles between components can go. 
-May not be needed but can be handy if the CSS becomes hard to track. */
