--- src/components/Button.js
@@ -1,1 +1,0 @@
-// Could create a custom button here.
