--- src/components/Card.js
@@ -1,2 +1,0 @@
-// Create the blueprint/structure for the card component here. 
-// The destination of the user's next aisle will be displayed in this card.
